requirements.txt
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
google-auth==2.6.4
google-auth-oauthlib==0.5.1
gspread==5.3.2
idna==3.3
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyTelegramBotAPI==4.4.0
python-dotenv==0.20.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
